Fail fast if trying to seed a non-demo database

+++ db/seeds.rb
@@ -3,6 +3,11 @@
 # This is the data used in the demo money tracker site, although there's nothing
 # to stop you using it in development too.
 
+unless ENV['DEMO_MODE']
+  puts "To avoid overwriting real data, seed data will only be loaded in demo mode"
+  exit
+end
+
 Transaction.destroy_all
 Account.destroy_all
